Cleanup to a build with no warnings (9.1, now 8.21)
The project's own code gave 96 warnings under -Wall -Wextra -Wshadow and
now gives none; those flags are on in the normal build, with Boost, Z3
and Graphviz included as SYSTEM so what is reported is ours.

Done from the list: createDomainDef's shadowed `name`, grapher.h's
unguarded node and silent "" node for a missing plan step, util.h's
unused code (including the self-seeding select_randomly overloads),
by-value in()/which_*/loader signatures, and get_subtasks' four copies
of one block. The loader's other copy-pasted blocks go too: effect
atoms, parameter lists, object lists, and the task-network code shared
by methods and the problem's :htn.

Three bugs found and fixed, each with a test:
- a problem whose :htn was an empty ordered network crashed the loader
  (exit 139): the earlier guard had gone into the method's copy only;
- a type used as a predicate in an effect made the loader index past
  an empty list (ptypes[pred][i]); lookups now fall back to __Object__;
- fol::Function printed "(fa b )" for "(f a b)".

No semantic change in the benchmark; sanitizers clean on all suites.

+++ lib/fol/Literal.h
@@ -23,7 +23,7 @@ namespace fol {
                                         const Literal<T>& lit) {
             out << "Literal(";
             out << "Predicate(" << lit.predicate << ") ";
-            for (int i=0; i < lit.args.size(); i++) {
+            for (size_t i=0; i < lit.args.size(); i++) {
                 out << lit.args.at(i);
                 if (i < lit.args.size() - 1) {
                     out << " ";
